% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pre_commit.R, R/githooks.R
\name{add_pre_commit}
\alias{add_pre_commit}
\alias{add_pre_push}
\title{Adds a git hook}
\usage{
add_pre_commit()

add_pre_push()

add_pre_commit()
}
\description{
Add a pre-commit or pre-push hook.
Both hooks are identical. They just get triggered at different times.

Adds a pre-commit hook
}
