% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all.R
\name{check_all}
\alias{check_all}
\title{Run all checks}
\usage{
check_all(
  pkg = NULL,
  lintr = NULL,
  namespace = NULL,
  r_filenames = NULL,
  version = NULL,
  gitignore = NULL,
  windows_issues = NULL,
  tidy_description = NULL,
  readme = NULL
)
}
\arguments{
\item{pkg, lintr, namespace, r_filenames, version, gitignore, windows_issues, readme, tidy_description}{Default \code{NULL}. See details for further information.}
}
\description{
This function will run all checks within the package.
See the corresponding check functions for details.
}
\details{
The arguments
for the function correspond to a particular check, e.g. check_ARGNAME(). By
default, all arguments are \code{NULL} and hence run. However, a value of
\code{FALSE} or changing the corresponding environment variable
\code{INTERGRATE_ARGNAME} can turn off the check.

Note: As this package matures, this function will include the newer checks.
}
