% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_windows_issues.R
\name{check_file_permissions}
\alias{check_file_permissions}
\title{Detects standard Windows related issues}
\usage{
check_file_permissions(repo_files = NULL, path = ".")
}
\arguments{
\item{repo_files}{By default, we use git to determine the files in the repo. By a
vector of files can be passed instead.}

\item{path}{Location of the package}
}
\description{
This check tests for windows line breaks and file permissions. It ensures
that the file is not executable (txt|md|Rmd|yml|json|).
}
\examples{
check_file_permissions()
}
