% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gitignore.R
\name{check_gitignore}
\alias{check_gitignore}
\title{Check gitignore file
Checks the gitignore files for recommended patterns. If any of these
patterns are missing, an error is raised.}
\usage{
check_gitignore()
}
\description{
Check gitignore file
Checks the gitignore files for recommended patterns. If any of these
patterns are missing, an error is raised.
}
\examples{
check_gitignore()
}
