% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lintr.R
\name{check_lintr}
\alias{check_lintr}
\title{The lintr check

Runs lint from the \code{lintr} package. Also scans for all
\code{.Rmd} and \code{.R} files in other directories.}
\usage{
check_lintr()
}
\description{
The lintr check

Runs lint from the \code{lintr} package. Also scans for all
\code{.Rmd} and \code{.R} files in other directories.
}
