% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta.R
\name{check_meta}
\alias{check_meta}
\title{Checks Meta files}
\usage{
check_meta()
}
\description{
Checks for a tidy description file (via the \code{usethis} function
\code{use_tidy_description}). It also checks that the version numbers conform
to the tidy format - X.Y.Z or X.Y.Z.9ABC, if the README is up to date, and
if .gitignore contains sensible values.
}
