% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_namespace.R
\name{check_namespace}
\alias{check_namespace}
\title{Namespace checks}
\usage{
check_namespace(no_imports = NULL)
}
\arguments{
\item{no_imports}{Default NULL. Number of package imports via \code{import()} allowed in the
NAMESPACE. If \code{NULL}, checks for environment variable NO_IMPORTS, otherwise
0.}
}
\description{
Importing too many packages can cause bugs due to namespace clashes. This check
enforces a set number of imports. As a rule of thumb, if you only use a few functions from
a package, use \code{importFrom} instead of \code{import} in your NAMESPACE file.
}
