% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_news.R
\name{check_news}
\alias{check_news}
\alias{check_all_news}
\title{Check NEWS format}
\usage{
check_news(pattern = NULL, path = ".")

check_all_news(pattern = NULL)
}
\arguments{
\item{pattern}{Regular expression NEWS title.}

\item{path}{Location of the package}
}
\description{
Checks that NEWS.md exists and title lines follow the correct format.
If the \code{pattern} is \code{NULL}, then the expected format is either:
\code{# pkg_name Version _YYYY-MM-DD_} or \code{# pkg_name (development version)}
}
\details{
The \code{check_all_news()} function isn't used in the CI. It's a
convient command line function to quickly assess your NEWS.md file
}
