% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg.R
\name{check_pkg}
\alias{check_pkg}
\title{Check R package}
\usage{
check_pkg(path = NULL)
}
\arguments{
\item{path}{Default \code{.} The location of the package}
}
\description{
Path is the CI project directory or \code{.}.
}
