% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_r_filenames.R
\name{check_r_filenames}
\alias{check_r_filenames}
\title{Check R/ filenames}
\usage{
check_r_filenames(extension = "R")
}
\arguments{
\item{extension}{Preferred R file extension. Default \code{R}.}
}
\description{
Checks that all filenames in the directory \code{R/}
are lower case and the file extension matches .\code{extension}
}
\examples{
check_r_filenames()
}
