% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_version.R
\name{check_version}
\alias{check_version}
\title{Version check}
\usage{
check_version(repo = "origin/master")
}
\arguments{
\item{repo}{Default origin/master. The repo to compare against.}
}
\description{
Check if the package version has been updated compared to the a
master repo (default is origin/master). This ensures that when branches are merged
the package version is bumped.
}
\details{
Files listed in .Rbuildignore don't count as changes.

Technically we only check for a change in the Version line of the DESCRIPTION file, not an
actual version increase.
}
