% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_windows_issues.R
\name{check_windows_issues}
\alias{check_windows_issues}
\title{Detects standard Windows related issues}
\usage{
check_windows_issues(permissions = TRUE, line_breaks = TRUE)
}
\arguments{
\item{permissions}{Default \code{TRUE}. Checks for file permissions.}

\item{line_breaks}{Default \code{TRUE}. Checks for Windows line breaks.}
}
\description{
This check tests for windows line breaks and file permissions. It ensures
that no file is public writeable or if it has a file extensions, (txt|md|Rmd|yml|json|R|r),
that it is not executable.
}
