% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tag.R
\name{create_tag}
\alias{create_tag}
\title{Auto-tagging via CI}
\usage{
create_tag(branch = "master")
}
\arguments{
\item{branch}{The branch where the tagging will occur. Default master.}
}
\description{
Automatically tag the commit via the version number. This requies the
environment variable GITHUB_TOKEN that has write permission.
}
