% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab_instead_of.R
\name{gitlab_instead_of}
\alias{gitlab_instead_of}
\title{Gitlab rewrite

Rewrites the gitlab URL. Using this means we don't require specific runner gitlab commands.}
\usage{
gitlab_instead_of()
}
\description{
Gitlab rewrite

Rewrites the gitlab URL. Using this means we don't require specific runner gitlab commands.
}
