% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_hooks.R
\name{add_pre_commit}
\alias{add_pre_commit}
\alias{add_pre_push}
\title{Adds a git hook}
\usage{
add_pre_commit()

add_pre_push()
}
\description{
Add a pre-commit or pre-push hook.
Both hooks are identical. They just get triggered at different times.
}
