% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_all.R
\name{check_all}
\alias{check_all}
\title{Run all checks}
\usage{
check_all(
  pkg = TRUE,
  lintr = TRUE,
  namespace = TRUE,
  r_filenames = TRUE,
  version = TRUE,
  gitignore = TRUE,
  tidy_description = TRUE,
  readme = TRUE,
  file_permissions = TRUE,
  line_breaks = TRUE,
  tag = TRUE,
  news = TRUE
)
}
\arguments{
\item{pkg}{Build and check package. See details for further information.}

\item{lintr, namespace, r_filenames, version, gitignore, readme, tidy_description}{Checks}

\item{file_permissions, line_breaks}{Windows related checks.}

\item{tag}{Create a tagged release.}

\item{news}{Check the top line of NEWS.md}
}
\description{
This function will run all checks within the package.
See the corresponding check functions for details.
}
\details{
The arguments for the function correspond to a particular check, e.g. check_ARGNAME().
Note: As this package matures, this function will include the newer checks.
}
