% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gitignore.R
\name{check_gitignore}
\alias{check_gitignore}
\title{Check gitignore file}
\usage{
check_gitignore(dir = ".")
}
\arguments{
\item{dir}{Default location for R files.}
}
\description{
Checks the gitignore files for recommended patterns. If any of these
patterns are missing, an error is raised.
}
\examples{
check_gitignore()
}
