% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_lintr.R
\name{check_lintr}
\alias{check_lintr}
\title{The lintr check

Runs lint from the \code{lintr} package.}
\usage{
check_lintr(readme = TRUE, vignettes = TRUE)
}
\arguments{
\item{readme}{Default \code{TRUE}. Should README.Rmd be checked}

\item{vignettes}{Default \code{TRUE}. Should the vignettes be linted.}
}
\description{
The lintr check

Runs lint from the \code{lintr} package.
}
