% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R, R/check_pkg.R
\name{get_pkg_tar_ball}
\alias{get_pkg_tar_ball}
\alias{check_pkg}
\title{Check R package}
\usage{
get_pkg_tar_ball()

check_pkg(path = NULL)
}
\arguments{
\item{path}{Default \code{.} The location of the package}
}
\description{
Path is the CI project directory or \code{.}.
}
