% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R, R/check_lintr.R, R/check_pkg.R,
%   R/check_r_filenames.R, R/check_tidy_description.R
\name{build_pkg}
\alias{build_pkg}
\alias{check_lintr}
\alias{check_pkg}
\alias{check_r_filenames}
\alias{check_tidy_description}
\title{Build R package}
\usage{
build_pkg(path = NULL)

check_lintr(readme = TRUE, vignettes = TRUE)

check_pkg(path = NULL, build = TRUE, install = TRUE)

check_r_filenames(extension = "R")

check_tidy_description(path = NULL)
}
\arguments{
\item{path}{Default \code{.} The location of the package}

\item{readme}{Default \code{TRUE}. Should README.Rmd be checked}

\item{vignettes}{Default \code{TRUE}. Should the vignettes be linted.}

\item{build}{Default \code{TRUE}. Call build_pkg() before checking.}

\item{install}{Default \code{TRUE}. Call install_pkg() after checking}

\item{extension}{Preferred R file extension. Default \code{R}.}
}
\description{
Path is the CI project directory or \code{.}.

Checks that all filenames are lower case and the file extension matches
.\code{extension}
}
