% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pkg.R
\name{check_pkg}
\alias{check_pkg}
\title{Check R package}
\usage{
check_pkg(path = ".")
}
\arguments{
\item{path}{Location of the package}
}
\description{
Path is the CI project directory or \code{.}.
Installs and checks the R package. The number of notes
and warnings is specified by enviromental variables.
Default number of Notes/Warnings is 0.
}
