% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_readme.R
\name{check_readme}
\alias{check_readme}
\title{Check the README.Rmd
If README.Rmd exists, the timestamp will be compared to README.md.
An error will be raised if README.md is a newer file.}
\usage{
check_readme()
}
\description{
Check the README.Rmd
If README.Rmd exists, the timestamp will be compared to README.md.
An error will be raised if README.md is a newer file.
}
\examples{
check_readme()
}
