% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tidy_description.R
\name{check_tidy_description}
\alias{check_tidy_description}
\title{Tidy description}
\usage{
check_tidy_description(path = NULL)
}
\arguments{
\item{path}{Default \code{.} The location of the package}
}
\description{
Checks for a tidy description file (via the \code{usethis} function
\code{use_tidy_description}). It also checks that the version numbers conform
to the tidy format - X.Y.Z or X.Y.Z.9ABC.
}
