% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_via_env.R
\name{check_via_env}
\alias{check_via_env}
\title{Run GITLAB checks via Environment variables

Looks for environment variables and runs associated checks.
This allows you to set environment variables at a project level, e.g. Gitlab.}
\usage{
check_via_env()
}
\description{
Run GITLAB checks via Environment variables

Looks for environment variables and runs associated checks.
This allows you to set environment variables at a project level, e.g. Gitlab.
}
