% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab_instead_of.R
\name{gitlab_instead_of}
\alias{gitlab_instead_of}
\title{Gitlab rewrite

Rewrites the gitlab URL. This can be handy if you want to push to another GitLab repo
on successful builds.}
\usage{
gitlab_instead_of()
}
\description{
Gitlab rewrite

Rewrites the gitlab URL. This can be handy if you want to push to another GitLab repo
on successful builds.
}
